+++ src/main/java/de/samply/file/bundle/PathsBundleManagerImpl.java
@@ -5,6 +5,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import org.springframework.util.FileSystemUtils;
 
 public class PathsBundleManagerImpl implements PathsBundleManager {
 
@@ -13,7 +14,7 @@ public class PathsBundleManagerImpl implements PathsBundleManager {
   /**
    * The current implementation assumes that there is only one bundle in the whole directory.
    *
-   * @param inputFolderPath  input folder path.
+   * @param inputFolderPath input folder path.
    */
   public PathsBundleManagerImpl(String inputFolderPath) {
     this.inputFolderPath = Paths.get(inputFolderPath);
@@ -22,7 +23,7 @@ public class PathsBundleManagerImpl implements PathsBundleManager {
   /**
    * The current implementation assumes that there is only one bundle in the whole directory.
    *
-   * @param inputFolderPath  input folder path.
+   * @param inputFolderPath input folder path.
    */
   public PathsBundleManagerImpl(Path inputFolderPath) {
     this.inputFolderPath = inputFolderPath;
@@ -70,10 +71,33 @@ public class PathsBundleManagerImpl implements PathsBundleManager {
       throws PathsBundleManagerException {
 
     if (pathsBundle != null) {
-      pathsBundle.applyToAllPaths(path -> pathsBundle.addPath(moveFileToOutputFolder(pathsBundle, path, outputFolderPath)));
+      Path oldDirectory = pathsBundle.getDirectory();
+      pathsBundle.applyToAllPaths(
+          path -> pathsBundle.addPath(moveFileToOutputFolder(pathsBundle, path, outputFolderPath)));
       pathsBundle.setDirectory(outputFolderPath);
+      removeSubdirectories(oldDirectory);
     }
 
+  }
+
+  private void removeSubdirectories(Path directory) throws PathsBundleManagerException {
+    try {
+      removeSubdirectoriesWithoutExceptionManagement(directory);
+    } catch (IOException | RuntimeException e) {
+      throw new PathsBundleManagerException(e);
+    }
+  }
+
+  private void removeSubdirectoriesWithoutExceptionManagement(Path directory) throws IOException {
+    if (Files.isDirectory(directory)) {
+      Files.list(directory).filter(Files::isDirectory).forEach(path -> {
+        try {
+          FileSystemUtils.deleteRecursively(path);
+        } catch (IOException e) {
+          throw new RuntimeException(e);
+        }
+      });
+    }
   }
 
   @Override
@@ -83,7 +107,8 @@ public class PathsBundleManagerImpl implements PathsBundleManager {
     PathsBundle pathsBundle2 = null;
     if (pathsBundle != null) {
       pathsBundle2 = pathsBundle.clone();
-      pathsBundle2.applyToAllPaths(path -> copyFileToOutputFolder(pathsBundle, path, outputFolderPath));
+      pathsBundle2.applyToAllPaths(
+          path -> copyFileToOutputFolder(pathsBundle, path, outputFolderPath));
       pathsBundle2.setDirectory(outputFolderPath);
     }
 
@@ -91,11 +116,12 @@ public class PathsBundleManagerImpl implements PathsBundleManager {
 
   }
 
-  protected Path moveFileToOutputFolder(PathsBundle pathsBundle, Path path, Path outputFolderPath) throws PathsBundleManagerException {
+  protected Path moveFileToOutputFolder(PathsBundle pathsBundle, Path path, Path outputFolderPath)
+      throws PathsBundleManagerException {
     try {
       Path tempPath = pathsBundle.getDirectory().relativize(path);
       Path tempPath2 = Paths.get(outputFolderPath.toString(), tempPath.toString());
-      if (!Files.exists(tempPath2.getParent())){
+      if (!Files.exists(tempPath2.getParent())) {
         Files.createDirectory(tempPath2.getParent());
       }
       Files.move(path, tempPath2, StandardCopyOption.REPLACE_EXISTING);
@@ -105,11 +131,12 @@ public class PathsBundleManagerImpl implements PathsBundleManager {
     }
   }
 
-  protected void copyFileToOutputFolder(PathsBundle pathsBundle, Path path, Path outputFolderPath) throws PathsBundleManagerException {
+  protected void copyFileToOutputFolder(PathsBundle pathsBundle, Path path, Path outputFolderPath)
+      throws PathsBundleManagerException {
     try {
       Path tempPath = pathsBundle.getDirectory().relativize(path);
       Path tempPath2 = Paths.get(outputFolderPath.toString(), tempPath.toString());
-      if (!Files.exists(tempPath2.getParent())){
+      if (!Files.exists(tempPath2.getParent())) {
         Files.createDirectory(tempPath2.getParent());
       }
       Files.copy(path, tempPath2, StandardCopyOption.REPLACE_EXISTING);