+++ src/main/java/de/samply/pseudonymisation/IdManagerResponse.java
@@ -1,0 +1,7 @@
+package de.samply.pseudonymisation;
+
+import java.util.List;
+
+public record IdManagerResponse (
+  List<PatientId> ids
+) {}