+++ src/main/java/de/samply/pseudonymisation/PatientId.java
@@ -1,0 +1,8 @@
+package de.samply.pseudonymisation;
+
+public record PatientId(
+    String idType,
+    String idString,
+    Boolean tentative,
+    String uri
+) {}